zerodha-dashboard-gemini/requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
kiteconnect==5.0.1
python-dotenv==1.0.1